3_AltayBajaURLS/requirements.txt
selenium>=4.0.0
openpyxl>=3.0.0
webdriver-manager>=3.8.0
pandas>=1.3.0
python-docx>=0.8.11
requests>=2.25.0

